Small improvements and tidy up

+++ BubblesDivePlanner/Commands/DiveStages/PreDiveStageAmbientPressure.cs
@@ -31,7 +31,7 @@ namespace BubblesDivePlanner.Commands.DiveStages
             var nitrogenFraction = 1.0 - _oxygenPercentage / 100.0 - _heliumPercentage / 100.0;
 
             //calculates ambient pressure
-            var pressureAmbient = 1.0 + (double)_depth / 10.0;
+            var pressureAmbient = 1.0 + _depth / 10.0;
 
             //calculates ambient pressure of each gas
             _diveProfile.PressureNitrogen = nitrogenFraction * pressureAmbient;
